+++ lib/src/type/platform/ios_type/cg_rect.dart
@@ -1,3 +1,20 @@
 import 'package:foundation_fluttify/foundation_fluttify.dart';
 
-class CGRect extends Ref {}
+class CGRect extends Ref {
+  static Future<CGRect> create(
+    double x,
+    double y,
+    double width,
+    double height,
+  ) async {
+    final refId = await kMethodChannel.invokeMethod('CGRect::create', {
+      'x': x,
+      'y': y,
+      'width': width,
+      'height': height,
+    });
+    return CGRect()
+      ..refId = refId
+      ..tag__ = 'platform';
+  }
+}