+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - enhance: [breaking change] 去除CoreLocation相关逻辑, 计划分离到core_location_fluttiy插件中去
 - feat: 增加android_view_View_OnTouchListener类; 增加ui_view.widget
 - feat: 增加av_audio_session_category_options
+- feat: CGRect增加create方法
 
 ## 0.6.1
 - enhance: android: *Handler的args参数使用Any类型, 内部再做转换